backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pillow
matplotlib
html2text
unstructured
langchain-unstructured
paddleocr
paddlenlp
dotenv
faiss-cpu
langgraph
langchain
langchain-openai
langchain-community
langchain-text-splitters
sqlalchemy
aiosqlite
aiomysql
requests
numpy